move script enums to proper location
