+++ ScriptAtRestServer/Enums/ScriptEnums.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ScriptAtRestServer.Services
+namespace ScriptAtRestServer.Enums
 {
     public class ScriptEnums
     {
